+++ app/settings.py
@@ -2,5 +2,6 @@ import os
 
 MONGO_URI = os.getenv('MONGO_URI')
 SECRET_KEY = os.getenv('SECRET_KEY')
-ANO = 21
+WERKZEUG_DEBUG_PIN = os.getenv('WERKZEUG_DEBUG_PIN')
+#ANO = 21
 #RANKING = os.getenv('RANKING','19-3')